PMA-207 - added staff time clock policy

+++ app/Http/Controllers/StaffTimeClockController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\StaffTimeClock\UpdateRequest;
 use App\Http\Resources\StaffTimeClockResource;
 use App\Http\Resources\StaffTimeClockResourceCollection;
 use App\Repositories\Contracts\StaffTimeClockRepository;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class StaffTimeClockController extends Controller
 {
@@ -31,9 +32,12 @@ class StaffTimeClockController extends Controller
      *
      * @param IndexRequest $request
      * @return StaffTimeClockResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', [StaffTimeClock::class, $request->get('propertyId')]);
+
         $staffTimeClocks = $this->staffTimeClockRepository->findBy($request->all());
 
         return new StaffTimeClockResourceCollection($staffTimeClocks);
@@ -42,11 +46,14 @@ class StaffTimeClockController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  StoreRequest $request
      * @return StaffTimeClockResource
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', [StaffTimeClock::class, $request->get('propertyId')]);
+
         $staffTimeClock = $this->staffTimeClockRepository->save($request->all());
 
         return new StaffTimeClockResource($staffTimeClock);
@@ -57,9 +64,12 @@ class StaffTimeClockController extends Controller
      *
      * @param StaffTimeClock $staffTimeClock
      * @return StaffTimeClockResource
+     * @throws AuthorizationException
      */
     public function show(StaffTimeClock $staffTimeClock)
     {
+        $this->authorize('show', $staffTimeClock);
+
         return new StaffTimeClockResource($staffTimeClock);
     }
 
@@ -69,9 +79,12 @@ class StaffTimeClockController extends Controller
      * @param UpdateRequest $request
      * @param StaffTimeClock $staffTimeClock
      * @return StaffTimeClockResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, StaffTimeClock $staffTimeClock)
     {
+        $this->authorize('update', $staffTimeClock);
+
         $staffTimeClock = $this->staffTimeClockRepository->update($staffTimeClock, $request->all());
 
         return new StaffTimeClockResource($staffTimeClock);
@@ -82,9 +95,12 @@ class StaffTimeClockController extends Controller
      *
      * @param StaffTimeClock $staffTimeClock
      * @return void
+     * @throws AuthorizationException
      */
     public function destroy(StaffTimeClock $staffTimeClock)
     {
+        $this->authorize('destroy', $staffTimeClock);
+
         $this->staffTimeClockRepository->delete($staffTimeClock);
 
         return response()->json(null, 204);
